Refactor: 프로젝트 구조 개선을 위해 패키지 추가 및 파일 이동 #27
