+++ app/src/main/java/com/example/hrr_android/LoadingActivity.kt
@@ -3,11 +3,9 @@ package com.example.hrr_android
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.example.hrr_android.databinding.ActivityLoadingBinding
+import com.example.hrr_android.onboarding.ui.ResultActivity
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
