+++ app/src/main/java/com/example/hrr_android/onboarding/ui/fragment/InfoSelectFragment.kt
@@ -1,10 +1,11 @@
-package com.example.hrr_android
+package com.example.hrr_android.onboarding.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.hrr_android.onboarding.utils.RadioGroupUtils
 import com.example.hrr_android.databinding.FragmentInfoSelectBinding
 
 class InfoSelectFragment : Fragment() {
