+++ app/src/main/res/layout/activity_intro.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
-    tools:context=".IntroActivity">
+    tools:context=".onboarding.ui.IntroActivity">
 
     <ImageView
         android:id="@+id/ellipse_background"
