+++ app/src/main/res/layout/activity_onboarding.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
-    tools:context=".OnboardingActivity">
+    tools:context=".onboarding.ui.OnboardingActivity">
 
     <!-- 뒤로가기 버튼 -->
     <ImageView
