+++ app/src/main/res/layout/fragment_goal.xml
@@ -5,7 +5,7 @@
     android:layout_height="wrap_content"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_marginStart="16dp"
-    tools:context=".GoalFragment">
+    tools:context=".onboarding.ui.fragment.GoalFragment">
 
     <!-- 제목 -->
     <TextView
