+++ app/src/main/res/layout/fragment_info_select.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:fillViewport="true"
     android:background="@color/white"
-    tools:context=".InfoSelectFragment">
+    tools:context=".onboarding.ui.fragment.InfoSelectFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
